+++ test/next/jdbc_test.clj
@@ -1,7 +1,7 @@
 ;; copyright (c) 2019 Sean Corfield, all rights reserved
 
 (ns next.jdbc-test
-  "Not exactly a test suite -- more a series of examples."
+  "Basic tests for the primary API of `next.jdbc`."
   (:require [clojure.string :as str]
             [clojure.test :refer [deftest is testing use-fixtures]]
             [next.jdbc :as jdbc]
@@ -11,7 +11,7 @@
             [next.jdbc.prepare :as prep]
             [next.jdbc.result-set :as rs]
             [next.jdbc.specs :as specs])
-  (:import (java.sql ResultSet ResultSetMetaData)))
+  (:import (java.sql ResultSet)))
 
 (set! *warn-on-reflection* true)
 
@@ -236,15 +236,3 @@ VALUES ('Pear', 'green', 49, 47)
               (into [] (map pr-str) (jdbc/plan (ds) ["select * from fruit"]))))
   (is (thrown? IllegalArgumentException
                (doall (take 3 (jdbc/plan (ds) ["select * from fruit"]))))))
-
-(deftest issue-73
-  ;; only postgresql cannot convert a java.util.Date to a timestamp column...
-  (try
-    (jdbc/execute-one! (ds) ["drop table temp_table"])
-    (catch Throwable _))
-  (jdbc/execute-one! (ds) ["create table temp_table (id int not null, deadline timestamp not null)"])
-  (try
-    (jdbc/execute-one! (ds) ["insert into temp_table (id, deadline) values (?,?)" 1 (java.util.Date.)])
-    (catch Throwable t
-      (println "Issue #73:" (db) "cannot convert java.util.Date to timestamp by default?")
-      (println (ex-message t)))))